refactor: rename folder __stories__ to stories
