+++ package.json
@@ -76,8 +76,8 @@
     "collectCoverageFrom": [
       "<rootDir>/src/**/*.{ts,tsx}",
       "!<rootDir>/src/**/*.stories.{ts,tsx}",
-      "!<rootDir>/src/**/__stories__/*",
-      "!<rootDir>/src/utils/tests/**"
+      "!<rootDir>/src/**/stories/**",
+      "!<rootDir>/src/**/tests/**"
     ]
   },
   "publishConfig": {
